+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix nested tags parsing.
+- Fixed comments being parsed as block, (tag name must now begin with an ASCII alpha).
 
 ## [0.1.0] - 2022-02-02